Creating client specific package and improving naming
