+++ workspaces/templates-lib/packages/template-user-management/src/codeChallenge.spec.ts
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import { getCodeChallenge } from './codeChallenge';
+import { getCodeChallenge } from './client/getCodeVerifier';
 
 it('Should generate code challenge', async () => {
   const challenge = await getCodeChallenge();
